// client-nachweise/src/features/nachweise/utils/nachweis-schema.ts
export interface Ausbilder {
  id: string;
  name: string;
  email: string;
  profileImageUrl: string | null;
}